init with an other architecture

+++ src/_chess/common.rs
@@ -1,6 +1,6 @@
 //! Common module
 
 pub(crate) struct Position {
-    x: i32,
-    y: i32,
+    x: i16,
+    y: i16,
 }
